rename execute_with_debugger_logger
