+++ packages/cli-core/src/package-manager-controller/execute_with_debugger_logger.test.ts
@@ -1,6 +1,6 @@
 import assert from 'assert';
 import { beforeEach, describe, it, mock } from 'node:test';
-import { executeWithDebugLogger } from './execute_with_logger.js';
+import { executeWithDebugLogger } from './execute_with_debugger_logger.js';
 
 const execaMock = mock.fn();
