+++ packages/cli-core/src/package-manager-controller/execute_with_debugger_logger.ts
